+++ ichnaea/api/locate/searcher.py
@@ -1,5 +1,5 @@
 """
-Abstract searcher and concrete country and position searchers each using
+Abstract searcher and concrete position and region searchers each using
 multiple sources to satisfy a given query.
 """
 
@@ -8,16 +8,16 @@ from collections import defaultdict
 from ichnaea.api.locate.cell import OCIDPositionSource
 from ichnaea.api.locate.fallback import FallbackPositionSource
 from ichnaea.api.locate.geoip import (
-    GeoIPCountrySource,
     GeoIPPositionSource,
+    GeoIPRegionSource,
 )
 from ichnaea.api.locate.internal import (
-    InternalCountrySource,
     InternalPositionSource,
+    InternalRegionSource,
 )
 from ichnaea.api.locate.result import (
-    Country,
     Position,
+    Region,
     ResultList,
 )
 from ichnaea.constants import DEGREE_DECIMAL_PLACES
@@ -30,17 +30,17 @@ def _configure_searcher(klass, settings, geoip_db=None, raven_client=None,
     return klass(settings, geoip_db, raven_client, redis_client, stats_client)
 
 
-def configure_country_searcher(settings, geoip_db=None, raven_client=None,
-                               redis_client=None, stats_client=None,
-                               _searcher=None):
+def configure_region_searcher(settings, geoip_db=None, raven_client=None,
+                              redis_client=None, stats_client=None,
+                              _searcher=None):
     """
     Configure and return a configured
-    :class:`~ichnaea.api.locate.searcher.CountrySearcher` instance.
+    :class:`~ichnaea.api.locate.searcher.RegionSearcher` instance.
 
     :param _searcher: Test-only hook to provide a pre-configured searcher.
     """
     return _configure_searcher(
-        CountrySearcher, settings, geoip_db=geoip_db,
+        RegionSearcher, settings, geoip_db=geoip_db,
         raven_client=raven_client, redis_client=redis_client,
         stats_client=stats_client, _searcher=_searcher)
 
@@ -147,21 +147,21 @@ class PositionSearcher(Searcher):
         return results.best()
 
 
-class CountrySearcher(Searcher):
+class RegionSearcher(Searcher):
     """
-    A CountrySearcher will return a country name and code.
+    A RegionSearcher will return a region name and code.
     """
 
-    result_type = Country
+    result_type = Region
     source_classes = (
-        ('internal', InternalCountrySource),
-        ('geoip', GeoIPCountrySource),
+        ('internal', InternalRegionSource),
+        ('geoip', GeoIPRegionSource),
     )
 
     def format_result(self, result):
         return {
-            'country_code': result.country_code,
-            'country_name': result.country_name,
+            'region_code': result.region_code,
+            'region_name': result.region_name,
             'fallback': result.fallback,
         }
 
@@ -176,7 +176,7 @@ class CountrySearcher(Searcher):
         # group by region code
         grouped = defaultdict(list)
         for result in found:
-            grouped[result.country_code].append(result)
+            grouped[result.region_code].append(result)
 
         regions = []
         for code, values in grouped.items():